overhaul the build and add boost
overview
--------

i overhauled the makefile to work w/ either gnu c++ 4.7 or 4.8. also
ripped out a bunch of the left-over stuff from where i copied it.

we need to do some non-trivial file I/O to read in all of the wsj data.
c++ stdandard libs for this are abysmal so i added [boost][boostfs].

unfortunately on os x this means we have to compile boost ourselves (see
below for an explanation). this won't be necessary on tacc.

building boost
--------------

    cd Downloads
    wget http://sourceforge.net/projects/boost/files/boost/1.55.0/boost_1_55_0.tar.gz/download
    mv download boost_1_55_0.tar.gz
    tar xzvf boost_1_55_0.tar.gz
    cd boost_1_55_0
    vim tools/build/v2/user-config.jam # or whatever editor you use
    # edit the part that says "Configure specific gcc version" to be:
    using gcc : 4.7 : /usr/local/bin/g++-4.7 ;
    # save and exit
    mkdir ~/boost
    ./bootstrap.sh --prefix=$HOME/boost
    ./b2 toolset=gcc-4.7
    ./b2 toolset=gcc-4.7 install

this should get you a copy in `~/boost` which the makefile expects when
using a `g++-4.X` compiler.

why we need to build boost from source
--------------------------------------

mac os x can't ship w/ gnu gcc b/c of licensing, so it's got it's own
llvm-based compiler. we need openmp for our project, which [isn't
currently supported by clang][clangomp] (the llvm c complier). so we
have to install gnu g++ via [brew][]. that's why we have to say
`g++-4.X` instead of just `g++`, since `g++` is aliased to `clang++`.

so if we `brew install boost`, there's no way to configure it to use
`g++-4.X` instead of `clang++`. the easiest solution i could find was
compiling boost, because the software engineer side of me HATES the
idea of writing directory iterators when perfectly good ones are in the
boost library.

[boostfs]: http://www.boost.org/doc/libs/1_55_0/libs/filesystem/doc/index.htm
[clangomp]: http://clang-omp.github.io
[brew]: http://brew.sh